refactor: 티켓 상세 화면 패키징 변경
