+++ app/src/main/java/woowacourse/movie/presentation/main/MainActivity.kt
@@ -13,7 +13,7 @@ import woowacourse.movie.databinding.ActivityMainBinding
 import woowacourse.movie.presentation.BaseActivity
 import woowacourse.movie.presentation.movies.MovieListFragment
 import woowacourse.movie.presentation.settings.SettingsFragment
-import woowacourse.movie.presentation.ticket.TicketListFragment
+import woowacourse.movie.presentation.ticket.list.TicketListFragment
 
 class MainActivity :
     BaseActivity<ActivityMainBinding>(R.layout.activity_main),
