--- app/src/main/java/woowacourse/movie/presentation/result/BookingResultPresenter.kt
@@ -1,12 +1,0 @@
-package woowacourse.movie.presentation.result
-
-import woowacourse.movie.domain.model.Ticket
-
-class BookingResultPresenter(
-    private val view: BookingResultContract.View,
-    private val ticket: Ticket,
-) : BookingResultContract.Presenter {
-    override fun loadBookingResult() {
-        view.showTicketInfo(ticket)
-    }
-}
