+++ app/src/main/java/woowacourse/movie/presentation/ticket/detail/TicketDetailPresenter.kt
@@ -1,0 +1,12 @@
+package woowacourse.movie.presentation.ticket.detail
+
+import woowacourse.movie.domain.model.Ticket
+
+class TicketDetailPresenter(
+    private val view: TicketDetailContract.View,
+    private val ticket: Ticket,
+) : TicketDetailContract.Presenter {
+    override fun loadBookingResult() {
+        view.showTicketInfo(ticket)
+    }
+}
