+++ app/src/main/java/woowacourse/movie/presentation/ticket/list/TicketListContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.presentation.ticket
+package woowacourse.movie.presentation.ticket.list
 
 import woowacourse.movie.domain.model.Ticket
