+++ app/src/main/java/woowacourse/movie/presentation/ticket/list/adapter/TicketViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.presentation.ticket.adapter
+package woowacourse.movie.presentation.ticket.list.adapter
 
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.common.adapter.ItemClickListener
