+++ app/src/main/res/layout/activity_ticket_detail.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".presentation.result.BookingResultActivity">
+    tools:context=".presentation.ticket.detail.TicketDetailActivity">
 
     <data>
