+++ app/src/main/res/layout/fragment_ticket_list.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".presentation.ticket.TicketListFragment">
+    tools:context=".presentation.ticket.list.TicketListFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/main"
